% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Combine levels of two factors}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{factor}

\item{b}{factor}
}
\value{
a factor with combined levels
}
\description{
This creates a new factor with the concatenated levels of the two input factors.
}
\examples{
fbind(as.factor(c("dog","cat")),as.factor(c("gerbil","parakeet")))
}
