% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackage21S001-package.R
\docType{package}
\name{testpackage21S001-package}
\alias{testpackage21S001}
\alias{testpackage21S001-package}
\title{testpackage21S001: Demo creating a new package for 413/613-001}
\description{
Not very much right now.
}
\author{
\strong{Maintainer}: Richard Ressler \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
